Fix Gen3 coin get/set

+++ PKHeX.Core/Saves/SAV3E.cs
@@ -110,8 +110,8 @@ namespace PKHeX.Core
 
         public override uint Coin
         {
-            get => BitConverter.ToUInt16(Large, 0x0494) ^ SecurityKey;
-            set => SetData(BitConverter.GetBytes(value ^ SecurityKey), 0x0494);
+            get => (ushort)(BitConverter.ToUInt16(Large, 0x0494) ^ SecurityKey);
+            set => SetData(BitConverter.GetBytes((ushort)(value ^ SecurityKey)), 0x0494);
         }
 
         private const int OFS_PCItem = 0x0498;
