+++ PKHeX.Core/Saves/SAV3FRLG.cs
@@ -100,8 +100,8 @@ namespace PKHeX.Core
 
         public override uint Coin
         {
-            get => BitConverter.ToUInt16(Large, 0x0294) ^ SecurityKey;
-            set => SetData(BitConverter.GetBytes(value ^ SecurityKey), 0x0294);
+            get => (ushort)(BitConverter.ToUInt16(Large, 0x0294) ^ SecurityKey);
+            set => SetData(BitConverter.GetBytes((ushort)(value ^ SecurityKey)), 0x0294);
         }
 
         private const int OFS_PCItem = 0x0298;
